Add `mongodb` package and new seeder
Also upgrade `mongoose` and add pnpm lock file.

Attempted to solve seed loader issue by upgrading `mongoose`, which I
thought may be part of the challenge, but after more than an hour
fighting it I decided to go with plain `mongodb` and create a new
seeder based on it.

Didn't remove `mongoose` package because I don't know if it may be
needed on next challenge.

Original seeder complained first about:

> DeprecationWarning: `open()` is deprecated in mongoose >= 4.11.0,
> use `openUri()` instead, or set the `useMongoClient` option if
> using `connect()` or `createConnection()`.
> See http://mongoosejs.com/docs/4.x/docs/connections.html#use-mongo-client

Which I solved, but then it throwed a second error:

> Error [MongoError]: Unsupported OP_QUERY command: insert. The client driver
> may require an upgrade.
> For more details see https://dochub.mongodb.org/core/legacy-opcode-removal

Current `mongoose` is using an ancient `mongodb` v2.2.34, which currently is
at v5.1.0, so I decided to switch to `mongodb` and solve the exercise.